% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_toolkit.R
\name{get_model}
\alias{get_model}
\title{Return the best model from a h2o grid search}
\usage{
get_model(data, target, algorithm, hyperparameters, loading = FALSE, ...)
}
\arguments{
\item{data}{The dataframe that is used for training and validating the
models.}

\item{target}{The name of the target variable.}

\item{algorithm}{A string that contains the name of the algorithm for the
models (gbm, randomForest, kmeans, glm, deeplearning, naivebayes, pca)}

\item{hyperparameters}{this is either a list of sequences for the values that
the grid search shall cover, or a list of concrete parameters}

\item{loading}{boolean value that is True when the function is used to
instantiate a new model with specific parameters}
}
\description{
The function passes further parameters, such as the search criteria, directly
to the h2o.grid() function. These parameters have to be explicitly named.
A new h2o instance is initialised. Pass startH2O=TRUE if that is not wanted.
}
\examples{
get_model(data, "duration", "gbm", hyper_parameter_list, search_criteria=search_criteria_list)
}
