% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchVisualization.R
\name{GridSearchGBM}
\alias{GridSearchGBM}
\title{A function that realizes and visualizes a GridSearch for Gradient Boosting Machines with H2O.}
\usage{
GridSearchGBM(train, test, predictorColumns, y, runtime)
}
\arguments{
\item{train}{The training frame as h2O Frame}

\item{test}{The validation frame as h2O frame}

\item{predictorColumns}{A vector of columns that are used for the prediction}

\item{y}{The goal variable}

\item{runtime}{Max time for model training}
}
\description{
This function allows you to further analyze GridSearches for GBM.
Therefore it initializes a default modell and compares it to the parameter values in a GridSearch
}
\examples{
GridSearchGBM(train, test, predictorColumns=c("column1", "column2", "column3"), y="columnY", 360)
}
