% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH2oFeatureImportance.R
\name{h2oViz.plotVariableImportance}
\alias{h2oViz.plotVariableImportance}
\title{A function that creates a styled h2o variable importance plot}
\usage{
h2oViz.plotVariableImportance(mod, threshold = 0.005)
}
\arguments{
\item{mod}{The model to query for feature importances}

\item{threshold}{Features with less relative importance than the threshold will be omitted (defaults to 0.005)}
}
\description{
A function that creates a styled h2o variable importance plot
}
\keyword{h2o}
