% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holiday.R
\name{is_holiday}
\alias{is_holiday}
\title{Return True when the given date is a holiday in the given state in Germany}
\usage{
is_holiday(day, state)
}
\arguments{
\item{day}{a date or String in the format yyyy-mm-dd}

\item{state}{a key for the state e.g. NI for Niedersachsen}
}
\description{
uses the https://feiertage-api.de/ api. Check the website for an overview of the
state keys
}
\examples{
is_holiday("2018-12-25", "NI")
}
