% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_toolkit.R
\name{load_model}
\alias{load_model}
\title{Loads a h2o model object}
\usage{
load_model(path, data)
}
\arguments{
\item{path}{A string with a correct file path}

\item{data}{data for instantiating the model}
}
\description{
The parameters from the serialized model object are read and a grid search
ist called with these parameters. To do this, data is needed to train the
model. Ideally, this is the same data with which the saved model was trained.
The new model object is returned.
}
\examples{
load_model("./trained_models/model_parameters_Test.rdata", cases_processed)

}
