% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_toolkit.R
\name{save_model}
\alias{save_model}
\title{Saves a h2o model object}
\usage{
save_model(model, path)
}
\arguments{
\item{model}{A h2o model object that shall be saved}

\item{path}{A string with a correct file path}
}
\description{
Serializes a h2o model object's parameters to the specified file path for
later reinstantiation. Currently the path needs to be a concrete file path
with the .rdata ending.
}
\examples{
save_model(gbm_model, "./trained_models/model_parameters_Test.rdata")

}
