% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_toolkit.R
\name{save_score}
\alias{save_score}
\title{Save information about a trained model in a csv file}
\usage{
save_score(model, data, path = "", split_seed = NA,
  custom_names = NULL, custom_values = NULL, custom_threshold = NULL)
}
\arguments{
\item{model}{a h2o model}

\item{data}{the data that was used for training and testing the model}

\item{path}{the path to where the csv shall be saved (optional)}

\item{split_seed}{if a seed was used to split the data in training and test set, this seed
can also be saved (optional)}

\item{custom_names}{custom variables can be added to the file. This variable
should contain the names of the columns as a list of strings (optional)}

\item{custom_values}{the values corresponding to the column names in custom_names. Lists as values are not supported! (optional)}

\item{custom_threshold}{the threshold that shall be used to determine the values of the different measures  (optional)}
}
\description{
Save information about a trained model in a csv file to compare model performance
over the lifetime of a running project. So far this function has only been tested for
binomial and regression models.
}
\examples{
save_score(gbm, checkedCases, custom_names=c("ersparnisProFall", "optimalThreshold"), custom_values=c(Optimal, optimalThreshold), custom_threshold = optimalThreshold)
}
