% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viadeeTheme.R
\name{viadeePlot}
\alias{viadeePlot}
\title{A function that styles a ggplot}
\usage{
viadeePlot(graph, backgroundGradientColors = c("#DFAD47", "#7EBCA9"))
}
\arguments{
\item{graph}{The graph to style.}

\item{backgroundGradientColors}{The two colors that form a gradient in the background}
}
\description{
This function allows you give your plots a viadee background
and some styling based on the economist theme.
Ideally this is the first thing you do with your graph
object, in order not to overpaint any content.
}
\examples{
viadeePlot(ggplot2::ggplot()) + ggplot2::geom_point()
}
\keyword{Background}
