% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viadeeTheme.R
\name{viadeePlot_fire}
\alias{viadeePlot_fire}
\title{A function that styles a ggplot}
\usage{
viadeePlot_fire(graph)
}
\arguments{
\item{graph}{The graph to style with a red-to-yellow background.}
}
\description{
This function allows you give your plots a viadee background
and some styling based on the economist theme.
Ideally this is the first thing you do with your graph
object, in order not to overpaint any content.
}
\examples{
viadeePlot_fire(ggplot2::ggplot()) + ggplot2::geom_point()
}
\keyword{Background}
